fix: remove EscapeHTML from Value
Sanitization of Value means that we can't set Texter to custom to code
like a `<svg>` icon for example. And also, it shouldn't be a security
issue unless you're using a Custom Texter that allows injection of the
user input in some way, and in those cases you can still call
`EscapeHTML` manually if needed.